// src/common/enums/product.enums.ts
export enum ProductCondition {
  NEW = 'new',
  LIKE_NEW = 'like_new',
}

export enum ProductAvailability {
  IN_STOCK = 'IN_STOCK',
  PREORDER = 'PREORDER',
  OUT_OF_STOCK = 'OUT_OF_STOCK',
}

export enum ProductStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
}

export enum GroupbuyType {
  NORMAL = 'normal',
  GROUPBUY = 'groupbuy',
  PREORDER = 'preorder',
}